possibly fixed image_index

+++ trunk/ENIGMAsystem/SHELL/Graphics_Systems/OpenGL/graphics_object.h
@@ -42,7 +42,7 @@ namespace enigma
   {
     //Sprites: these are mostly for higher tiers...
       int sprite_index;
-      int image_index;
+      double image_index;
       double image_speed;
       
       int image_number; //read only
